Add files via upload
Changed the project root directory from WFR_Templates to Wils_FoolsRoad.

Added additional files named WFR_Fools_Road_Landscape_and_Geo.umap and WilsFoolsRoad.uasset. The %geo.umap file is a blank level in reference to the original map. the other asset is just the outer-level ID data asset.

With these combined and moving templates inside the main project, this allows the geo map to stay loaded and in reference for the layers created from the templates. Later releases I plan on making an easier reference for the changes, right now the default Fools Road geo map is loaded in the templates for general reference along with the pre-placed POI capture zones. Make sure to remove this when you are done making any changes.